Update shake screen & timer
